refactor(dex): remove Dex module and related controllers, services, and entities to streamline the codebase and improve maintainability
