+++ src/plugins/dex/config/dex-contracts.config.ts
