+++ src/plugins/dex/entities/pair-summary.entity.ts
