+++ src/plugins/dex/services/pair-summary.service.ts
