+++ src/plugins/dex/services/pair.service.ts
